+++ packages/react-router/lib/router/router.ts
@@ -384,9 +384,7 @@ export type HydrationState = Partial<
 /**
  * Future flags to toggle new feature behavior
  */
-export interface FutureConfig {
-  unstable_middleware: boolean;
-}
+export interface FutureConfig {}
 
 /**
  * Initialization options for createRouter
@@ -435,7 +433,12 @@ export interface StaticHandler {
       skipRevalidation?: boolean;
       dataStrategy?: DataStrategyFunction<unknown>;
       unstable_generateMiddlewareResponse?: (
-        query: (r: Request) => Promise<StaticHandlerContext | Response>,
+        query: (
+          r: Request,
+          args?: {
+            filterMatchesToLoad?: (match: AgnosticDataRouteMatch) => boolean;
+          },
+        ) => Promise<StaticHandlerContext | Response>,
       ) => MaybePromise<Response>;
     },
   ): Promise<StaticHandlerContext | Response>;
@@ -867,11 +870,13 @@ export function createRouter(init: RouterInit): Router {
   );
   let inFlightDataRoutes: AgnosticDataRouteObject[] | undefined;
   let basename = init.basename || "/";
+  if (!basename.startsWith("/")) {
+    basename = `/${basename}`;
+  }
   let dataStrategyImpl = init.dataStrategy || defaultDataStrategyWithMiddleware;
 
   // Config driven behavior flags
   let future: FutureConfig = {
-    unstable_middleware: false,
     ...init.future,
   };
   // Cleanup function for history
@@ -1802,6 +1807,20 @@ export function createRouter(init: RouterInit): Router {
       if (discoverResult.type === "aborted") {
         return { shortCircuited: true };
       } else if (discoverResult.type === "error") {
+        if (discoverResult.partialMatches.length === 0) {
+          let { matches, route } = getShortCircuitMatches(dataRoutes);
+          return {
+            matches,
+            pendingActionResult: [
+              route.id,
+              {
+                type: ResultType.error,
+                error: discoverResult.error,
+              },
+            ],
+          };
+        }
+
         let boundaryId = findNearestBoundary(discoverResult.partialMatches)
           .route.id;
         return {
@@ -1996,6 +2015,17 @@ export function createRouter(init: RouterInit): Router {
       if (discoverResult.type === "aborted") {
         return { shortCircuited: true };
       } else if (discoverResult.type === "error") {
+        if (discoverResult.partialMatches.length === 0) {
+          let { matches, route } = getShortCircuitMatches(dataRoutes);
+          return {
+            matches,
+            loaderData: {},
+            errors: {
+              [route.id]: discoverResult.error,
+            },
+          };
+        }
+
         let boundaryId = findNearestBoundary(discoverResult.partialMatches)
           .route.id;
         return {
@@ -3621,7 +3651,14 @@ export function createStaticHandler(
           },
           async () => {
             let res = await generateMiddlewareResponse(
-              async (revalidationRequest: Request) => {
+              async (
+                revalidationRequest: Request,
+                opts: {
+                  filterMatchesToLoad?:
+                    | ((match: AgnosticDataRouteMatch) => boolean)
+                    | undefined;
+                } = {},
+              ) => {
                 let result = await queryImpl(
                   revalidationRequest,
                   location,
@@ -3630,7 +3667,9 @@ export function createStaticHandler(
                   dataStrategy || null,
                   skipLoaderErrorBubbling === true,
                   null,
-                  filterMatchesToLoad || null,
+                  "filterMatchesToLoad" in opts
+                    ? (opts.filterMatchesToLoad ?? null)
+                    : null,
                   skipRevalidation === true,
                 );